* feature/plugin/isplugindeveloper:
  利用 s="1" 來對語言包的 index 來執行 parsetag
  顯示語言包的 index 名稱 - part.4
  顯示語言包的 index 名稱 - part.3
  顯示語言包的 index 名稱 - part.2
  顯示語言包的 index 名稱 - part.1
  textarea 增加當取得焦點時，執行 onkeyup(); 調整高度
  修改 copyright 為 textarea 方便輸入
  增加允許修改版權訊息
  update doc url
  bugfix - part.3
  bugfix - part.2 修正造成插件語言包無法緩存的問題
  bugfix - part.1
  插件開發者模式 1=開啟 2=開啟(並且開啟顯示模板的 <hook> 嵌入點)

+++ discuzx/upload/extensions/hooks/hooks_core.php
@@ -160,10 +160,14 @@ function _eFunc_cachedata_Before_get_syscache($_EVENT, $conf) {
 		updatecache($cachenames);
 		*/
 		foreach ($cachenames as $k) {
-			if(!isset($_del_cache[$k])) {
-				$_del_cache[$k] = true;
+			if(!isset($_del_cache[$k])
+				// bugfix 修正造成插件語言包無法緩存的問題
+				&& substr($k, 0, 6) != 'plugin'
+			) {
 				DB::query("DELETE FROM ".DB::table('common_syscache')." WHERE cname = '$k' LIMIT 1");
 			}
+
+			$_del_cache[$k] = true;
 		}
 	}
 }
